Fix 64bit WKO parsing GPS data
The decoding of GPS data erroneously assumed that a
signed long was 32 bits, which is not true on a 64bit
platform.

Fixes #333.

+++ src/WkoRideFile.cpp
@@ -374,7 +374,7 @@ WKO_UCHAR *WkoParseRawData(WKO_UCHAR *fb, WKO_ULONG version, RideFile *rideFile,
                         break;
                     case 'G' : /* two longs */
                         {
-                            signed long llat, llon;
+                            int32_t llat, llon;
                             char slat[20], slon[20];
 
                             // stored 2s complement
